Start the project anew with `symfony new`.
Limaju stands for Liquid Majority Judgment.
The goal of this project is to provide a headless API
for liquid majority judgment polling as a service.

Framadate will also do this eventually,
but we need an algorithmic constitution
that anyone can read and contribute to.

This project is therefore story-driven,
and extensively (ab)uses Gherkin.

The feature files will be in another repository,
and we'll use git submodules to include them here.

Also, sorry for doing this in PHP.
We tried doing this in Ruby,
but Ruby's regex support is sadly inadequate,
no-one wants to fix it (named captures are broken),
and we're not brave enough to fix it ourselves.
That being said, we'd welcome a server implementation in Ruby,
so long as it implements the same features.