feat: add an Interface for Ranking (the new Tallier) algorithms
This architecture will enable us to add multiple ranking algorithms,
each with their own set of options and options constraints,
since options values will come from userland and may not be trusted.

We'll probably use the Form component, or perhaps just the Validation.
We can probably use a Documenter to generate the API documentation
for each algorithm, from the annotations in the options class.