docs: explain why we want to support multiple ranking algorithms
Even if the so-called `Usual Judgment` is less elegant,
it's fair to leave room for it (data! data! empiricism!).

It will also (more importantly) leave room for a
`Liquid Majority Judgment` algorithm as well,
even though the delegation graph probably ought to be decentralized
in the long run, for obvious privacy reasons, and safety.