refacto: keep the controllers short and sweet
This is a stepping stone to the next public tool endpoint:
tally resolution

Note that we're using the DataTransformer component but not the Form.
(it makes little sense to use it here, but perhaps I'm wrong)

This is more in-line with how we are supposed to handle parameters,
and it will help hiding the (future) tally integrity logic,
or default grade configuration propagation up to userland.